feat: create and add register anchor component
