+++ src/app/(pages)/(auth)/login/page.tsx
@@ -1,42 +1,14 @@
-export default function LoginForm() {
+import LoginForm from "./components/loginForm";
+import RegisterAnchor from "./components/registerAnchor";
+
+export default function LoginPage() {
   return (
-    <div className="flex items-center justify-center min-h-screen bg-[#C08B74] p-4">
+    <div className="flex flex-col items-center justify-center min-h-screen bg-[#C08B74] p-4">
       <div className="bg-white p-8 rounded-lg shadow-md w-full max-w-md">
         <h2 className="text-2xl font-bold mb-6 text-center">Login</h2>
-        <form>
-          <div className="mb-4">
-            <label className="block text-sm font-medium mb-2" htmlFor="email">
-              Email
-            </label>
-            <input
-              type="email"
-              id="email"
-              className="w-full p-2 border border-gray-300 rounded"
-              required
-            />
-          </div>
-          <div className="mb-4">
-            <label
-              className="block text-sm font-medium mb-2"
-              htmlFor="password"
-            >
-              Senha
-            </label>
-            <input
-              type="password"
-              id="password"
-              className="w-full p-2 border border-gray-300 rounded"
-              required
-            />
-          </div>
-          <button
-            type="submit"
-            className="w-full bg-[#D8671E] text-white py-2 rounded hover:bg-[#C0581A] transition duration-200"
-          >
-            Entrar
-          </button>
-        </form>
+        <LoginForm />
       </div>
+      <RegisterAnchor />
     </div>
   );
 }
